got rid of separate .hpp files with layer declarations; instead, put everything into the respective .cpp files
