+++ modules/dnn/include/opencv2/dnn/all_layers.hpp
@@ -72,12 +72,18 @@ namespace dnn
   - Dropout (since it does nothing on forward pass -))
 */
 
+    class CV_EXPORTS BlankLayer : public Layer
+    {
+    public:
+        static Ptr<BlankLayer> create(const LayerParams &params);
+    };
+
     //! LSTM recurrent layer
-    class CV_EXPORTS_W LSTMLayer : public Layer
+    class CV_EXPORTS LSTMLayer : public Layer
     {
     public:
         /** Creates instance of LSTM layer */
-        static CV_WRAP Ptr<LSTMLayer> create();
+        static Ptr<LSTMLayer> create();
 
         /** Set trained weights for LSTM layer.
         LSTM behavior on each step is defined by current input, previous output, previous cell state and learned weights.
@@ -109,27 +115,27 @@ namespace dnn
         @param Wx is matrix defining how current input is transformed to internal gates (i.e. according to abovemtioned notation is @f$ W_x @f$)
         @param b  is bias vector (i.e. according to abovemtioned notation is @f$ b @f$)
         */
-        CV_WRAP virtual void setWeights(const Mat &Wh, const Mat &Wx, const Mat &b) = 0;
+        virtual void setWeights(const Mat &Wh, const Mat &Wx, const Mat &b) = 0;
 
         /** @brief Specifies shape of output blob which will be [[`T`], `N`] + @p outTailShape.
           * @details If this parameter is empty or unset then @p outTailShape = [`Wh`.size(0)] will be used,
           * where `Wh` is parameter from setWeights().
           */
-        CV_WRAP virtual void setOutShape(const std::vector<int> &outTailShape = std::vector<int>()) = 0;
+        virtual void setOutShape(const std::vector<int> &outTailShape = std::vector<int>()) = 0;
 
         /** @brief Set @f$ h_{t-1} @f$ value that will be used in next forward() calls.
           * @details By-default @f$ h_{t-1} @f$ is inited by zeros and updated after each forward() call.
           */
-        CV_WRAP virtual void setH(const Mat &H) = 0;
+        virtual void setH(const Mat &H) = 0;
         /** @brief Returns current @f$ h_{t-1} @f$ value (deep copy). */
-        CV_WRAP virtual Mat getH() const = 0;
+        virtual Mat getH() const = 0;
 
         /** @brief Set @f$ c_{t-1} @f$ value that will be used in next forward() calls.
           * @details By-default @f$ c_{t-1} @f$ is inited by zeros and updated after each forward() call.
           */
-        CV_WRAP virtual void setC(const Mat &C) = 0;
+        virtual void setC(const Mat &C) = 0;
         /** @brief Returns current @f$ c_{t-1} @f$ value (deep copy). */
-        CV_WRAP virtual Mat getC() const = 0;
+        virtual Mat getC() const = 0;
 
         /** @brief Specifies either interpet first dimension of input blob as timestamp dimenion either as sample.
           *
@@ -139,12 +145,12 @@ namespace dnn
           * If flag is set to false then shape of input blob will be interpeted as [`N`, `[data dims]`].
           * In this case each forward() call will make one iteration and produce one timestamp with shape [`N`, `[out dims]`].
           */
-        CV_WRAP virtual void setUseTimstampsDim(bool use = true) = 0;
+        virtual void setUseTimstampsDim(bool use = true) = 0;
 
         /** @brief If this flag is set to true then layer will produce @f$ c_t @f$ as second output.
          * @details Shape of the second output is the same as first output.
          */
-        CV_WRAP virtual void setProduceCellOutput(bool produce = false) = 0;
+        virtual void setProduceCellOutput(bool produce = false) = 0;
 
         /** In common case it use single input with @f$x_t@f$ values to compute output(s) @f$h_t@f$ (and @f$c_t@f$).
          * @param input should contain packed values @f$x_t@f$
@@ -164,11 +170,11 @@ namespace dnn
     };
 
     //! Classical recurrent layer
-    class CV_EXPORTS_W RNNLayer : public Layer
+    class CV_EXPORTS RNNLayer : public Layer
     {
     public:
         /** Creates instance of RNNLayer */
-        static CV_WRAP Ptr<RNNLayer> create();
+        static Ptr<RNNLayer> create();
 
         /** Setups learned weights.
 
@@ -184,12 +190,12 @@ namespace dnn
         @param Who is @f$ W_{xo} @f$ matrix
         @param bo  is @f$ b_{o}  @f$ vector
         */
-        CV_WRAP virtual void setWeights(const Mat &Wxh, const Mat &bh, const Mat &Whh, const Mat &Who, const Mat &bo) = 0;
+        virtual void setWeights(const Mat &Wxh, const Mat &bh, const Mat &Whh, const Mat &Who, const Mat &bo) = 0;
 
         /** @brief If this flag is set to true then layer will produce @f$ h_t @f$ as second output.
          * @details Shape of the second output is the same as first output.
          */
-        CV_WRAP virtual void setProduceHiddenOutput(bool produce = false) = 0;
+        virtual void setProduceHiddenOutput(bool produce = false) = 0;
 
         /** Accepts two inputs @f$x_t@f$ and @f$h_{t-1}@f$ and compute two outputs @f$o_t@f$ and @f$h_t@f$.
 
@@ -205,29 +211,29 @@ namespace dnn
         void forward(std::vector<Mat*> &input, std::vector<Mat> &output);
     };
 
-    class CV_EXPORTS_W BaseConvolutionLayer : public Layer
+    class CV_EXPORTS BaseConvolutionLayer : public Layer
     {
     public:
 
-        CV_PROP_RW Size kernel, stride, pad, dilation, adjustPad;
-        CV_PROP_RW String padMode;
+        Size kernel, stride, pad, dilation, adjustPad;
+        String padMode;
     };
 
-    class CV_EXPORTS_W ConvolutionLayer : public BaseConvolutionLayer
+    class CV_EXPORTS ConvolutionLayer : public BaseConvolutionLayer
     {
     public:
 
-        static CV_WRAP Ptr<BaseConvolutionLayer> create(Size kernel = Size(3, 3), Size stride = Size(1, 1), Size pad = Size(0, 0), Size dilation = Size(1, 1));
+        static Ptr<BaseConvolutionLayer> create(Size kernel = Size(3, 3), Size stride = Size(1, 1), Size pad = Size(0, 0), Size dilation = Size(1, 1));
     };
 
-    class CV_EXPORTS_W DeconvolutionLayer : public BaseConvolutionLayer
+    class CV_EXPORTS DeconvolutionLayer : public BaseConvolutionLayer
     {
     public:
 
-        static CV_WRAP Ptr<BaseConvolutionLayer> create(Size kernel = Size(3, 3), Size stride = Size(1, 1), Size pad = Size(0, 0), Size dilation = Size(1, 1), Size adjustPad = Size());
+        static Ptr<BaseConvolutionLayer> create(Size kernel = Size(3, 3), Size stride = Size(1, 1), Size pad = Size(0, 0), Size dilation = Size(1, 1), Size adjustPad = Size());
     };
 
-    class CV_EXPORTS_W LRNLayer : public Layer
+    class CV_EXPORTS LRNLayer : public Layer
     {
     public:
 
@@ -236,18 +242,18 @@ namespace dnn
             CHANNEL_NRM,
             SPATIAL_NRM
         };
-        CV_PROP_RW int type;
+        int type;
 
-        CV_PROP_RW int size;
-        CV_PROP_RW double alpha, beta, bias;
-        CV_PROP_RW bool normBySize;
+        int size;
+        double alpha, beta, bias;
+        bool normBySize;
 
-        static CV_WRAP Ptr<LRNLayer> create(int type = LRNLayer::CHANNEL_NRM, int size = 5,
+        static Ptr<LRNLayer> create(int type = LRNLayer::CHANNEL_NRM, int size = 5,
                                             double alpha = 1, double beta = 0.75, double bias = 1,
                                             bool normBySize = true);
     };
 
-    class CV_EXPORTS_W PoolingLayer : public Layer
+    class CV_EXPORTS PoolingLayer : public Layer
     {
     public:
 
@@ -258,139 +264,157 @@ namespace dnn
             STOCHASTIC
         };
 
-        CV_PROP_RW int type;
-        CV_PROP_RW Size kernel, stride, pad;
-        CV_PROP_RW bool globalPooling;
-        CV_PROP_RW String padMode;
+        int type;
+        Size kernel, stride, pad;
+        bool globalPooling;
+        String padMode;
 
-        static CV_WRAP Ptr<PoolingLayer> create(int type = PoolingLayer::MAX, Size kernel = Size(2, 2),
+        static Ptr<PoolingLayer> create(int type = PoolingLayer::MAX, Size kernel = Size(2, 2),
                                                 Size stride = Size(1, 1), Size pad = Size(0, 0),
                                                 const cv::String& padMode = "");
-        static CV_WRAP Ptr<PoolingLayer> createGlobal(int type = PoolingLayer::MAX);
+        static Ptr<PoolingLayer> createGlobal(int type = PoolingLayer::MAX);
     };
 
-    class CV_EXPORTS_W SoftmaxLayer : public Layer
+    class CV_EXPORTS SoftmaxLayer : public Layer
     {
     public:
 
-        static CV_WRAP Ptr<SoftmaxLayer> create(int axis = 1);
+        static Ptr<SoftmaxLayer> create(int axis = 1);
     };
 
-    class CV_EXPORTS_W InnerProductLayer : public Layer
-    {
-    public:
-        CV_PROP_RW int axis;
-
-        static CV_WRAP Ptr<InnerProductLayer> create(int axis = 1);
-    };
-
-    class CV_EXPORTS_W MVNLayer : public Layer
-    {
-    public:
-        CV_PROP_RW double eps;
-        CV_PROP_RW bool normVariance, acrossChannels;
-
-        static CV_WRAP Ptr<MVNLayer> create(bool normVariance = true, bool acrossChannels = false, double eps = 1e-9);
-    };
-
-    /* Reshaping */
-
-    class CV_EXPORTS_W ReshapeLayer : public Layer
-    {
-    public:
-        CV_PROP_RW std::vector<int> newShapeDesc;
-        CV_PROP_RW Range newShapeRange;
-
-        static CV_WRAP Ptr<ReshapeLayer> create(const std::vector<int> &newShape, Range applyingRange = Range::all(),
-                                                bool enableReordering = false);
-    };
-
-    class CV_EXPORTS_W ConcatLayer : public Layer
+    class CV_EXPORTS InnerProductLayer : public Layer
     {
     public:
         int axis;
 
-        static CV_WRAP Ptr<ConcatLayer> create(int axis = 1);
+        static Ptr<InnerProductLayer> create(int axis = 1);
     };
 
-    class CV_EXPORTS_W SplitLayer : public Layer
+    class CV_EXPORTS MVNLayer : public Layer
+    {
+    public:
+        double eps;
+        bool normVariance, acrossChannels;
+
+        static Ptr<MVNLayer> create(bool normVariance = true, bool acrossChannels = false, double eps = 1e-9);
+    };
+
+    /* Reshaping */
+
+    class CV_EXPORTS ReshapeLayer : public Layer
+    {
+    public:
+        std::vector<int> newShapeDesc;
+        Range newShapeRange;
+
+        static Ptr<ReshapeLayer> create(const std::vector<int> &newShape, Range applyingRange = Range::all(),
+                                                bool enableReordering = false);
+    };
+
+    class CV_EXPORTS FlattenLayer : public Layer
+    {
+    public:
+        static Ptr<FlattenLayer> create(LayerParams &params);
+    };
+
+    class CV_EXPORTS ConcatLayer : public Layer
+    {
+    public:
+        int axis;
+
+        static Ptr<ConcatLayer> create(int axis = 1);
+    };
+
+    class CV_EXPORTS SplitLayer : public Layer
     {
     public:
         int outputsCount; //!< Number of copies that will be produced (is ignored when negative).
 
-        static CV_WRAP Ptr<SplitLayer> create(int outputsCount = -1);
+        static Ptr<SplitLayer> create(int outputsCount = -1);
     };
 
-    class CV_EXPORTS_W SliceLayer : public Layer
+    class CV_EXPORTS SliceLayer : public Layer
     {
     public:
-        CV_PROP_RW int axis;
-        CV_PROP std::vector<int> sliceIndices;
+        int axis;
+        std::vector<int> sliceIndices;
 
-        static CV_WRAP Ptr<SliceLayer> create(int axis);
-        static CV_WRAP Ptr<SliceLayer> create(int axis, const std::vector<int> &sliceIndices);
+        static Ptr<SliceLayer> create(int axis);
+        static Ptr<SliceLayer> create(int axis, const std::vector<int> &sliceIndices);
+    };
+
+    class CV_EXPORTS PermuteLayer : public Layer
+    {
+    public:
+        static Ptr<PermuteLayer> create(const LayerParams& params);
+    };
+
+    class CV_EXPORTS PaddingLayer : public Layer
+    {
+    public:
+        static Ptr<PaddingLayer> create(const LayerParams& params);
     };
 
     /* Activations */
 
-    class CV_EXPORTS_W ReLULayer : public Layer
+    class CV_EXPORTS ReLULayer : public Layer
     {
     public:
-        CV_PROP_RW double negativeSlope;
+        double negativeSlope;
 
-        static CV_WRAP Ptr<ReLULayer> create(double negativeSlope = 0);
+        static Ptr<ReLULayer> create(double negativeSlope = 0);
     };
 
-    class CV_EXPORTS_W ChannelsPReLULayer : public Layer
+    class CV_EXPORTS ChannelsPReLULayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<ChannelsPReLULayer> create();
+        static Ptr<ChannelsPReLULayer> create();
     };
 
-    class CV_EXPORTS_W TanHLayer : public Layer
+    class CV_EXPORTS TanHLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<TanHLayer> create();
+        static Ptr<TanHLayer> create();
     };
 
-    class CV_EXPORTS_W SigmoidLayer : public Layer
+    class CV_EXPORTS SigmoidLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<SigmoidLayer> create();
+        static Ptr<SigmoidLayer> create();
     };
 
-    class CV_EXPORTS_W BNLLLayer : public Layer
+    class CV_EXPORTS BNLLLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<BNLLLayer> create();
+        static Ptr<BNLLLayer> create();
     };
 
-    class CV_EXPORTS_W AbsLayer : public Layer
+    class CV_EXPORTS AbsLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<AbsLayer> create();
+        static Ptr<AbsLayer> create();
     };
 
-    class CV_EXPORTS_W PowerLayer : public Layer
+    class CV_EXPORTS PowerLayer : public Layer
     {
     public:
-        CV_PROP_RW double power, scale, shift;
+        double power, scale, shift;
 
-        static CV_WRAP Ptr<PowerLayer> create(double power = 1, double scale = 1, double shift = 0);
+        static Ptr<PowerLayer> create(double power = 1, double scale = 1, double shift = 0);
     };
 
     /* Layers using in semantic segmentation */
 
-    class CV_EXPORTS_W CropLayer : public Layer
+    class CV_EXPORTS CropLayer : public Layer
     {
     public:
-        CV_PROP int startAxis;
-        CV_PROP std::vector<int> offset;
+        int startAxis;
+        std::vector<int> offset;
 
         static Ptr<CropLayer> create(int start_axis, const std::vector<int> &offset);
     };
 
-    class CV_EXPORTS_W EltwiseLayer : public Layer
+    class CV_EXPORTS EltwiseLayer : public Layer
     {
     public:
         enum EltwiseOp
@@ -403,22 +427,46 @@ namespace dnn
         static Ptr<EltwiseLayer> create(EltwiseOp op, const std::vector<int> &coeffs);
     };
 
-    class CV_EXPORTS_W BatchNormLayer : public Layer
+    class CV_EXPORTS BatchNormLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<BatchNormLayer> create(bool hasWeights, bool hasBias, float epsilon);
+        static Ptr<BatchNormLayer> create(bool hasWeights, bool hasBias, float epsilon);
     };
 
-    class CV_EXPORTS_W MaxUnpoolLayer : public Layer
+    class CV_EXPORTS MaxUnpoolLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<MaxUnpoolLayer> create(Size poolKernel, Size poolPad, Size poolStride);
+        static Ptr<MaxUnpoolLayer> create(Size poolKernel, Size poolPad, Size poolStride);
     };
 
-    class CV_EXPORTS_W ScaleLayer : public Layer
+    class CV_EXPORTS ScaleLayer : public Layer
     {
     public:
-        static CV_WRAP Ptr<ScaleLayer> create(bool hasBias);
+        static Ptr<ScaleLayer> create(bool hasBias);
+    };
+
+    class CV_EXPORTS ShiftLayer : public Layer
+    {
+    public:
+        static Ptr<ShiftLayer> create(const LayerParams& params);
+    };
+
+    class CV_EXPORTS PriorBoxLayer : public Layer
+    {
+    public:
+        static Ptr<PriorBoxLayer> create(const LayerParams& params);
+    };
+
+    class CV_EXPORTS DetectionOutputLayer : public Layer
+    {
+    public:
+        static Ptr<DetectionOutputLayer> create(const LayerParams& params);
+    };
+
+    class NormalizeBBoxLayer : public Layer
+    {
+    public:
+        static Ptr<NormalizeBBoxLayer> create(const LayerParams& params);
     };
 
 //! @}
