+++ modules/dnn/samples/fcn_semsegm.cpp
@@ -1,7 +1,6 @@
 #include <opencv2/dnn.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/core/ocl.hpp>
 using namespace cv;
 using namespace cv::dnn;
