+++ modules/dnn/samples/torch_enet.cpp
@@ -5,7 +5,6 @@ Sample of using OpenCV dnn module with Torch ENet model.
 #include <opencv2/dnn.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/core/ocl.hpp>
 using namespace cv;
 using namespace cv::dnn;
