+++ modules/dnn/src/layers/flatten_layer.cpp
@@ -41,7 +41,6 @@
 
 #include "../precomp.hpp"
 #include "layers_common.hpp"
-#include "flatten_layer.hpp"
 #include <float.h>
 #include <algorithm>
 
@@ -50,58 +49,73 @@ namespace cv
 namespace dnn
 {
 
-FlattenLayer::FlattenLayer(LayerParams &params) : Layer(params)
+class FlattenLayerImpl : public FlattenLayer
 {
-    _startAxis = params.get<int>("axis", 1);
-    _endAxis = params.get<int>("end_axis", -1);
+public:
+    FlattenLayerImpl(LayerParams &params)
+    {
+        _startAxis = params.get<int>("axis", 1);
+        _endAxis = params.get<int>("end_axis", -1);
+        setParamsFrom(params);
+    }
+    void allocate(const std::vector<Mat*> &inputs, std::vector<Mat> &outputs){
+        checkInputs(inputs);
+        const Mat& inp0 = *inputs[0];
+
+        _numAxes = inp0.dims;
+        _endAxis = _endAxis < 0 ? _endAxis + _numAxes : _endAxis;
+        CV_Assert(_startAxis >= 0);
+        CV_Assert(_endAxis >= _startAxis && _endAxis < (int)_numAxes);
+
+        size_t flattenedDimensionSize = inp0.total(_startAxis, _endAxis+1);
+
+        resultShape.clear();
+        for (int i = 0; i < _startAxis; i++)
+        {
+            resultShape.push_back(inp0.size[i]);
+        }
+        resultShape.push_back(flattenedDimensionSize);
+        for (size_t i = _endAxis + 1; i < _numAxes; i++)
+        {
+            resultShape.push_back(inp0.size[i]);
+        }
+        CV_Assert(resultShape.size() <= 4);
+
+        for (size_t i = 0; i < inputs.size(); i++)
+        {
+            //in-place
+            outputs[i] = inputs[i]->reshape(1, (int)resultShape.size(), &resultShape[0]);
+        }
+    }
+
+    void forward(std::vector<Mat*> &inputs, std::vector<Mat> &outputs)
+    {
+        for (size_t i = 0; i < inputs.size(); i++)
+        {
+            outputs[i] = inputs[i]->reshape(1, (int)resultShape.size(), &resultShape[0]);
+        }
+    }
+
+    void checkInputs(const std::vector<Mat*> &inputs)
+    {
+        CV_Assert(inputs.size() > 0);
+        for (size_t i = 1; i < inputs.size(); i++)
+        {
+            CV_Assert(inputs[i]->size == inputs[0]->size);
+        }
+    }
+
+    int _startAxis;
+    int _endAxis;
+    size_t _numAxes;
+
+    std::vector<int> resultShape;
+};
+
+Ptr<FlattenLayer> FlattenLayer::create(LayerParams& params)
+{
+    return Ptr<FlattenLayer>(new FlattenLayerImpl(params));
 }
 
-void FlattenLayer::checkInputs(const std::vector<Mat*> &inputs)
-{
-    CV_Assert(inputs.size() > 0);
-    for (size_t i = 1; i < inputs.size(); i++)
-    {
-        CV_Assert(inputs[i]->size == inputs[0]->size);
-    }
-}
-
-void FlattenLayer::allocate(const std::vector<Mat*> &inputs, std::vector<Mat> &outputs)
-{
-    checkInputs(inputs);
-    const Mat& inp0 = *inputs[0];
-
-    _numAxes = inp0.dims;
-    _endAxis = _endAxis < 0 ? _endAxis + _numAxes : _endAxis;
-    CV_Assert(_startAxis >= 0);
-    CV_Assert(_endAxis >= _startAxis && _endAxis < (int)_numAxes);
-
-    size_t flattenedDimensionSize = inp0.total(_startAxis, _endAxis+1);
-
-    resultShape.clear();
-    for (int i = 0; i < _startAxis; i++)
-    {
-        resultShape.push_back(inp0.size[i]);
-    }
-    resultShape.push_back(flattenedDimensionSize);
-    for (size_t i = _endAxis + 1; i < _numAxes; i++)
-    {
-        resultShape.push_back(inp0.size[i]);
-    }
-    CV_Assert(resultShape.size() <= 4);
-
-    for (size_t i = 0; i < inputs.size(); i++)
-    {
-        //in-place
-        outputs[i] = inputs[i]->reshape(1, (int)resultShape.size(), &resultShape[0]);
-    }
-}
-
-void FlattenLayer::forward(std::vector<Mat*> &inputs, std::vector<Mat> &outputs)
-{
-    for (size_t i = 0; i < inputs.size(); i++)
-    {
-        outputs[i] = inputs[i]->reshape(1, (int)resultShape.size(), &resultShape[0]);
-    }
-}
 }
 }
