+++ modules/dnn/src/layers/recurrent_layers.cpp
@@ -40,7 +40,6 @@
 //M*/
 
 #include "../precomp.hpp"
-#include "recurrent_layers.hpp"
 #include "op_blas.hpp"
 #include <iostream>
 #include <iterator>
