+++ modules/dnn/src/precomp.hpp
@@ -42,3 +42,4 @@
 #include <opencv2/core.hpp>
 #include "cvconfig.h"
 #include <opencv2/dnn.hpp>
+#include <opencv2/dnn/all_layers.hpp>
